Formatting fixes

+++ gibthon/unafold.py
@@ -14,6 +14,7 @@ class UnaFolder():
 		self.safety = safety
 		self.na_salt = na_salt
 		self.mg_salt = mg_salt
+		self.warnings = []
 		
 	def hybrid_options(self):
 		return ' -n DNA -t %.2f -T %.2f -N %.2f -M %.2f --mfold=5,-1,100 ' %(self.t + self.safety, self.t + self.safety, self.na_salt, self.mg_salt)
@@ -25,9 +26,7 @@ class UnaFolder():
 			print e
 			return False
 		stdout, stderr = p.communicate()
-		dir(p)
 		if p.returncode > 0:
-			print("oh return code")
 			return False
 		return True
 	
@@ -37,12 +36,11 @@ class UnaFolder():
 		w = open(self.wd + self.name, 'w')
 		w.write(str(sequence))
 		w.close()
-		devnull = open(os.devnull,'w')
 		cline = self.una_root + 'hybrid-ss-min' + self.hybrid_options() + self.wd + self.name
 		if not self.process(cline):
 			print('Could not hybridise')
 			return False
-		cline = self.una_root + 'boxplot_ng -t "Energy dotplot "' + self.wd + self.name + '.plot'
+		cline = self.una_root + 'boxplot_ng -t "Energy dotplot " ' + self.wd + self.name + '.plot'
 		if not self.process(cline):
 			print('Could not convert boxplot')
 			return False
@@ -54,12 +52,46 @@ class UnaFolder():
 		ss = csv.DictReader(csvfile, delimiter='\t')
 		warnings = []
 		for r in ss:
-			if int(r['j']) == len(self.seq()):
-				warnings.append(r['length'], float(r['energy']/10))
+			if int(r['j']) == len(sequence):
+				self.warnings.append((r['length'], float(r['energy'])/10))
 		cline = 'convert ' + self.wd + self.name + '.ps ' + self.wd + self.name + '.png'
 		if not self.process(cline):
 			return False
 		for f in os.listdir(self.wd):
-			if os.path.isfile(f) and f.startswith(self.name):
+			if os.path.isfile(f) and f.startswith(self.name) and not f.endswith('.png'):
 				os.remove(f)
 		os.chdir(cwd)
+		return True, self.wd + self.name + '.png'
+		
+	def mis_prime(self, target, primer):
+		primerloc = self.name + '-primer'
+		targetloc = self.name + '-target'
+		cwd = os.getcwd()
+		os.chdir(self.wd)
+		w = open(self.wd + primerloc, 'w')
+		w.write(str(primer))
+		w.close
+		w = open(self.wd + targetloc, 'w')
+		w.write(str(target))
+		w.close()
+		cline = self.una_root + 'hybrid-min' + self.hybrid_options() + self.wd + targetloc + ' ' + self.wd + primerloc
+		if not self.process(cline):
+			print('Could not hybridise')
+			return False
+		try:
+			csvfile = open(self.wd + targetloc + '-' + primerloc + '.plot')
+		except IOError as e:
+			print e
+			return False
+		ss = csv.DictReader(csvfile, delimiter='\t')
+		for r in ss:
+			l = int(r['length'])
+			j = len(primer) - (int(r['j']) - len(target))
+			i = (int(r['i']) + l - 1)
+			if (j == 0 and i == len(target)) or (l == 1 or l == len(primer)-1):
+				continue
+			else:
+				self.warnings.append((l,j,i,float(r['energy'])/10))
+		
+		os.chdir(cwd)
+		return True
